fix survey modal
