+++ packages/main-ui/src/pages/surveys/new/page.rs
@@ -62,7 +62,7 @@ pub fn SurveyCreatePage(lang: Language, survey_id: Option<i64>) -> Element {
                     combination_error: ctrl.combination_error(),
 
                     clicked_complete_button: move |_| async move {
-                        ctrl.open_start_project_modal(lang).await;
+                        ctrl.clicked_start_project().await;
                     },
 
                     change_selected_tab: move |selected: bool| {
